api_service/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
alembic==1.10.3
python-dotenv==1.0.0
redis==4.5.4
httpx==0.24.0
pydantic[email]==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
